requirements.txt
dbt-core==1.9.4
dbt-postgres==1.9.0
pandas
sqlalchemy
psycopg2-binary